# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Generic SNAP/ASPI-compliant payment gateway API: B2B access\ntoken issuance, admin client onboarding, signature-generation\nutilities, and Virtual Account (transfer-VA) transaction and\nmerchant-dashboard endpoints.",
        "title": "SNAP Payment Integration Gateway API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/openapi/v1.0/transfer-va/create-va": {
            "post": {
                "security": [
                    {
                        "SnapTimestamp": []
                    },
                    {
                        "SnapSignature": []
                    }
                ],
                "description": "Merchant-initiated upsert of a Virtual Account (ASPI VAUpsertRequest). This performs a real state-changing action: it creates or updates a persistent Virtual Account record.",
                "tags": [
                    "Merchant VA Dashboard"
                ],
                "summary": "Create or update a Virtual Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request timestamp, ISO 8601",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Symmetric signature; compute via POST /api/v1/utilities/signature-service",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique external ID for this request",
                        "name": "X-EXTERNAL-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique key for this request; enforced by IdempotencyMiddleware. A repeated key with an identical payload replays the cached response; a repeated key with a different payload is rejected with 422.",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "VA create/update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantCreateVARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantCreateVAResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Field Format / Invalid Mandatory Field / missing Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantCreateVAResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (mapped from downstream error)",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantCreateVAResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict: request already in progress for this Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantCreateVAResponse"
                        }
                    },
                    "422": {
                        "description": "Idempotency-Key reused with a different payload",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantCreateVAResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantCreateVAResponse"
                        }
                    }
                }
            }
        },
        "/openapi/v1.0/transfer-va/delete-va": {
            "delete": {
                "security": [
                    {
                        "SnapTimestamp": []
                    },
                    {
                        "SnapSignature": []
                    }
                ],
                "description": "Merchant-initiated deletion of a Virtual Account (ASPI DeleteVARequest). This performs a real state-changing action: it permanently removes/deactivates the Virtual Account record.",
                "tags": [
                    "Merchant VA Dashboard"
                ],
                "summary": "Delete a Virtual Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request timestamp, ISO 8601",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Symmetric signature; compute via POST /api/v1/utilities/signature-service",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique external ID for this request",
                        "name": "X-EXTERNAL-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique key for this request; enforced by IdempotencyMiddleware. A repeated key with an identical payload replays the cached response; a repeated key with a different payload is rejected with 422.",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "VA delete request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantDeleteVARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantDeleteVAResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Field Format / Invalid Mandatory Field / missing Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantDeleteVAResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (mapped from downstream error)",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantDeleteVAResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict: request already in progress for this Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantDeleteVAResponse"
                        }
                    },
                    "422": {
                        "description": "Idempotency-Key reused with a different payload",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantDeleteVAResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantDeleteVAResponse"
                        }
                    }
                }
            }
        },
...
        "/openapi/v1.0/transfer-va/list": {
            "post": {
                "security": [
                    {
                        "SnapTimestamp": []
                    },
                    {
                        "SnapSignature": []
                    }
                ],
                "description": "Merchant-initiated paginated listing of Virtual Account transactions, filterable by date range, status and VA number. Read-only.",
                "tags": [
                    "Merchant VA Dashboard"
                ],
                "summary": "List Virtual Account transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request timestamp, ISO 8601",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Symmetric signature; compute via POST /api/v1/utilities/signature-service",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique external ID for this request",
                        "name": "X-EXTERNAL-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique key for this request; enforced by IdempotencyMiddleware since ListVA is called via POST. A repeated key with an identical payload replays the cached response; a repeated key with a different payload is rejected with 422.",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "VA list filter/pagination request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantListVARequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantListVAResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Field Format / missing Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantListVAResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (mapped from downstream error)",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantListVAResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict: request already in progress for this Idempotency-Key",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantListVAResponse"
                        }
                    },
                    "422": {
                        "description": "Idempotency-Key reused with a different payload",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantListVAResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.MerchantListVAResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Amount": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "domain.BilingualText": {
            "type": "object",
            "properties": {
                "english": {
                    "type": "string"
                },
                "indonesia": {
                    "type": "string"
                }
            }
        },
        "domain.BillDetail": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "billAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "billAmountLabel": {
                    "type": "string"
                },
                "billAmountValue": {
                    "type": "string"
                },
                "billCode": {
                    "type": "string"
                },
                "billDescription": {
                    "$ref": "#/definitions/domain.BilingualText"
                },
                "billName": {
                    "type": "string"
                },
                "billNo": {
                    "type": "string"
                },
                "billReferenceNo": {
                    "type": "string"
                },
                "billShortName": {
                    "type": "string"
                },
                "billSubCompany": {
                    "type": "string"
                },
                "billerReferenceId": {
                    "type": "string"
                },
                "reason": {
                    "$ref": "#/definitions/domain.BilingualText"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "domain.MerchantCreateVARequest": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "description": "AdditionalInfo carries proprietary extensions per ASPI VAUpsertRequest's\nadditionalInfo.dbUrlProcess slot (aspi-open-api-va.yaml:317-320) — the\nmerchant payment callback URL is passed here as additionalInfo.dbUrlProcess,\nnot as a top-level field (there is no such field in the spec).",
                    "type": "object",
                    "additionalProperties": true
                },
                "billDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BillDetail"
                    }
                },
                "customerNo": {
                    "type": "string"
                },
                "expiredDate": {
                    "type": "string"
                },
                "feeAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "freeTexts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BilingualText"
                    }
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "totalAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "trxId": {
                    "type": "string"
                },
                "virtualAccountEmail": {
                    "type": "string"
                },
                "virtualAccountName": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                },
                "virtualAccountPhone": {
                    "type": "string"
                },
                "virtualAccountTrxType": {
                    "type": "string"
                }
            }
        },
        "domain.MerchantCreateVAResponse": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "virtualAccountData": {
                    "$ref": "#/definitions/domain.MerchantVAData"
                }
            }
        },
        "domain.MerchantDeleteVAData": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "customerNo": {
                    "type": "string"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "trxId": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                }
            }
        },
        "domain.MerchantDeleteVARequest": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "customerNo": {
                    "type": "string"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "trxId": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                }
            }
        },
        "domain.MerchantDeleteVAResponse": {
            "type": "object",
            "properties": {
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                },
                "virtualAccountData": {
                    "$ref": "#/definitions/domain.MerchantDeleteVAData"
                }
            }
        },
        "domain.MerchantListVARequest": {
            "type": "object",
            "properties": {
                "fromDate": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "toDate": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                }
            }
        },
        "domain.MerchantListVAResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.VAListItem"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/domain.Pagination"
                },
                "responseCode": {
                    "type": "string"
                },
                "responseMessage": {
                    "type": "string"
                }
            }
        },
        "domain.MerchantVAData": {
            "type": "object",
            "properties": {
                "additionalInfo": {
                    "type": "object",
                    "additionalProperties": true
                },
                "billDetails": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BillDetail"
                    }
                },
                "customerNo": {
                    "type": "string"
                },
                "expiredDate": {
                    "type": "string"
                },
                "feeAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "freeTexts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.BilingualText"
                    }
                },
                "lastUpdateDate": {
                    "type": "string"
                },
                "partnerServiceId": {
                    "type": "string"
                },
                "paymentDate": {
                    "type": "string"
                },
                "totalAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "trxId": {
                    "type": "string"
                },
                "virtualAccountEmail": {
                    "type": "string"
                },
                "virtualAccountName": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                },
                "virtualAccountPhone": {
                    "type": "string"
                },
                "virtualAccountTrxType": {
                    "type": "string"
                }
            }
        },
        "domain.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                },
                "totalRows": {
                    "type": "integer"
                }
            }
        },
...
        "domain.VAListItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "customerName": {
                    "type": "string"
                },
                "customerNo": {
                    "type": "string"
                },
                "expiredDate": {
                    "type": "string"
                },
                "paidAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "status": {
                    "type": "string"
                },
                "totalAmount": {
                    "$ref": "#/definitions/domain.Amount"
                },
                "transactionDate": {
                    "type": "string"
                },
                "virtualAccountNo": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminToken": {
            "description": "Static admin API key configured via the ADMIN_API_KEY environment variable. Required for all /admin/* endpoints.",
            "type": "apiKey",
            "name": "X-Admin-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Bearer access token issued by POST /openapi/v1.0/access-token/b2b, e.g. \"Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SnapClientKey": {
            "description": "Client identifier issued by the admin during onboarding. Required on the B2B access-token endpoint.",
            "type": "apiKey",
            "name": "X-CLIENT-KEY",
            "in": "header"
        },
        "SnapSignature": {
            "description": "Request signature. Compute via POST /api/v1/utilities/signature-auth (asymmetric, access-token endpoint) or POST /api/v1/utilities/signature-service (symmetric, transaction endpoints).",
            "type": "apiKey",
            "name": "X-SIGNATURE",
            "in": "header"
        },
        "SnapTimestamp": {
            "description": "Request timestamp in ISO 8601 format (e.g. 2026-07-24T10:00:00+07:00).",
            "type": "apiKey",
            "name": "X-TIMESTAMP",
            "in": "header"
        }
    }
}